requirements.txt
pyserial
mido
python-rtmidi